requirements.txt
rapidfuzz==3.13.0
nltk==3.9.2
hydra-core==1.3.2
pandas==2.3.3
requests==2.31.0
omegaconf==2.3.0
pyarrow==21.0.0
fastparquet==2024.11.0
python-Levenshtein==0.27.1